finalize cleanup closes #301
